DRTII-1412 Separate features into v1 & v2. Persist version
