+++ jvm/src/main/scala/uk/gov/homeoffice/drt/protobuf/serialisation/ModelAndFeaturesConversion.scala
@@ -1,11 +1,13 @@
 package uk.gov.homeoffice.drt.protobuf.serialisation
 
 import org.slf4j.LoggerFactory
-import uk.gov.homeoffice.drt.prediction.arrival.FeatureColumns.{OneToMany, Single}
+import uk.gov.homeoffice.drt.prediction.arrival.features.{FeatureColumnsV1, FeatureColumnsV2, OneToManyFeature, SingleFeature}
+import uk.gov.homeoffice.drt.prediction.arrival.features.FeatureColumnsV1.{OneToMany, Single}
 import uk.gov.homeoffice.drt.prediction.{FeaturesWithOneToManyValues, ModelAndFeatures, RegressionModel}
 import uk.gov.homeoffice.drt.protobuf.messages.ModelAndFeatures._
 import uk.gov.homeoffice.drt.time.{LocalDate, SDate, SDateLike}
 
+import scala.collection.View.Single
 import scala.util.{Failure, Success, Try}
 
 object ModelAndFeaturesConversion {
@@ -22,7 +24,8 @@ object ModelAndFeaturesConversion {
                                     (implicit sdate: Long => SDateLike,
                                      sdateFromLocalDate: LocalDate => SDateLike): Option[ModelAndFeatures] = {
     val model = msg.model.map(modelFromMessage).getOrElse(throw new Exception("No value for model"))
-    val features = msg.features.map(featuresFromMessage).getOrElse(throw new Exception("No value for features"))
+    val featuresVersion = msg.featuresVersion.getOrElse(1)
+    val features = msg.features.map(f => featuresFromMessage(f, featuresVersion)).getOrElse(throw new Exception("No value for features"))
     val targetName = msg.targetName.getOrElse(throw new Exception("Mandatory parameter 'targetName' not specified"))
     val examplesTrainedOn = msg.examplesTrainedOn.getOrElse(throw new Exception("Mandatory parameter 'examplesTrainedOn' not specified"))
     val improvementPct = msg.improvementPct.getOrElse(throw new Exception("Mandatory parameter 'improvement' not specified"))
@@ -38,16 +41,17 @@ object ModelAndFeaturesConversion {
   def modelFromMessage(msg: RegressionModelMessage): RegressionModel =
     RegressionModel(msg.coefficients, msg.intercept.getOrElse(throw new Exception("No value for intercept")))
 
-  def featuresFromMessage(msg: FeaturesMessage)
+  def featuresFromMessage(msg: FeaturesMessage, version: Int)
                          (implicit
                           sdateFromLong: Long => SDateLike,
                           sdateFromLocalDate: LocalDate => SDateLike,
                          ): FeaturesWithOneToManyValues = {
-    implicit val now: () => SDate.JodaSDate = () => SDate.now()
-    val singles = msg.singleFeatures
-      .map(l => Try(Single.fromLabel(l)))
-    val oneToManys = msg.oneToManyFeatures
-      .map(l => Try(OneToMany.fromLabel(l)))
+    val (singleFromLabel, oneToManyFromLabel) = version match {
+      case 1 => (FeatureColumnsV1.Single.fromLabel _, FeatureColumnsV1.OneToMany.fromLabel _)
+      case 2 => (FeatureColumnsV2.Single.fromLabel _, FeatureColumnsV2.OneToMany.fromLabel _)
+    }
+    val singles = msg.singleFeatures.map(l => Try(singleFromLabel(l)))
+    val oneToManys = msg.oneToManyFeatures.map(l => Try(oneToManyFromLabel(l)))
 
     val allFeatures = (oneToManys ++ singles).collect { case Success(f) => f }
 
@@ -63,10 +67,10 @@ object ModelAndFeaturesConversion {
   def featuresToMessage(features: FeaturesWithOneToManyValues): FeaturesMessage = {
     FeaturesMessage(
       oneToManyFeatures = features.features.collect {
-        case feature: OneToMany[_] => feature.label
+        case feature: OneToManyFeature[_] => feature.label
       },
       singleFeatures = features.features.collect {
-        case feature: Single[_] => feature.label
+        case feature: SingleFeature[_] => feature.label
       },
       oneToManyValues = features.oneToManyValues
     )
@@ -80,6 +84,7 @@ object ModelAndFeaturesConversion {
       examplesTrainedOn = Option(modelAndFeatures.examplesTrainedOn),
       improvementPct = Option(modelAndFeatures.improvementPct),
       timestamp = Option(now),
+      featuresVersion = Option(modelAndFeatures.featuresVersion),
     )
   }
