+++ shared/src/main/scala/uk/gov/homeoffice/drt/prediction/FeaturesWithOneToManyValues.scala
@@ -1,14 +1,14 @@
 package uk.gov.homeoffice.drt.prediction
 
-import uk.gov.homeoffice.drt.prediction.arrival.FeatureColumns.{Feature, OneToMany, Single}
+import uk.gov.homeoffice.drt.prediction.arrival.features.{Feature, OneToManyFeature, SingleFeature}
 
 
 case class FeaturesWithOneToManyValues(features: List[Feature[_]], oneToManyValues: IndexedSeq[String]) {
-  def oneToManyFeatures: List[OneToMany[_]] = features.collect {
-    case otm: OneToMany[_] => otm
+  def oneToManyFeatures: List[OneToManyFeature[_]] = features.collect {
+    case otm: OneToManyFeature[_] => otm
   }
 
-  def singleFeatures: Seq[Single[_]] = features.collect {
-    case otm: Single[_] => otm
+  def singleFeatures: Seq[SingleFeature[_]] = features.collect {
+    case otm: SingleFeature[_] => otm
   }
 }
