+++ shared/src/main/scala/uk/gov/homeoffice/drt/prediction/arrival/ToChoxModelAndFeatures.scala
@@ -1,15 +1,16 @@
 package uk.gov.homeoffice.drt.prediction.arrival
 
-import uk.gov.homeoffice.drt.arrivals.Arrival
 import uk.gov.homeoffice.drt.prediction.{FeaturesWithOneToManyValues, RegressionModel}
 
 object ToChoxModelAndFeatures {
   val targetName: String = "to-chox"
+  val featuresVersion: Int = 1
 }
 
 case class ToChoxModelAndFeatures(model: RegressionModel,
                                   features: FeaturesWithOneToManyValues,
                                   examplesTrainedOn: Int,
                                   improvementPct: Double) extends ArrivalModelAndFeatures {
+  override val featuresVersion: Int = ToChoxModelAndFeatures.featuresVersion
   override val targetName: String = ToChoxModelAndFeatures.targetName
 }
